Add sorting and refactor paging (#37)
* change dto schema

* add authors into the book table

* add trigger for authors_books table

* implement mapstruct for Dtos

* Change scripts for DB initialize and change sortable component

* move logic of calculation pageable parameters to frontend part

+++ BookLibrary-cdi/backend/src/main/java/com/softserveinc/booklibrary/backend/entity/Author.java
@@ -1,5 +1,6 @@
 package com.softserveinc.booklibrary.backend.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 
@@ -15,9 +16,6 @@ import javax.persistence.Table;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.Fetch;
-import org.hibernate.annotations.FetchMode;
 
 @Getter
 @Setter
@@ -44,7 +42,7 @@ public class Author implements AbstractEntity<Integer> {
 	private LocalDateTime createDate;
 
 	@Column(name = "author_rating")
-	private Double authorRating;
+	private BigDecimal authorRating;
 
 	@ManyToMany
 	@JoinTable(
